+++ lesson10-web-app-modules-and-tests/domain/src/main/kotlin/pt/isel/domain/Game/money/Ante.kt
@@ -4,7 +4,7 @@ import java.time.Instant
 
 data class Ante(
     val amount: Int,
-    val matchId: Long,
+    val matchId: Int,
     val roundNumber: Int,
     val createdAt: Instant = Instant.now()
 ) {
@@ -16,13 +16,13 @@ data class Ante(
     }
 
     // Helper to collect ante from a player's wallet into the pot.
-    fun collectBet(userId: Long, pot: Pot, wallet: Wallet): Pair<Pot, Wallet> {
+    fun collectBet(userId: Int, pot: Pot, wallet: Wallet): Pair<Pot, Wallet> {
         require(userId > 0L) { "userId inválido" }
         require(!pot.closed) { "Pote encerrado" }
         require(pot.matchId == matchId && pot.roundNumber == roundNumber) {
             "Ante não corresponde ao pote (match/round)"
         }
-        require(wallet.userId.toLong() == userId) { "Wallet não corresponde ao user" }
+        require(wallet.userId == userId) { "Wallet não corresponde ao user" }
 
         val updatedWallet = wallet.withdraw(amount)
         val updatedPot = pot.addContribution(userId, amount)
